backend/pyproject.toml
[project]
name = "pro-mcp"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "dotenv>=0.9.9",
    "fastmcp>=2.12.3",
    "python-jose>=3.5.0",
    "sqlalchemy>=2.0.43",
    "stytch>=13.22.0",
]
